application/requirements.txt
boto3==1.20.31
botocore==1.23.31
python-dotenv==0.19.2